RubyLintBear: Move to Ruby directory
