+++ bears/tests/ruby/RubyLintBearTest.py
@@ -6,7 +6,7 @@ from queue import Queue
 sys.path.insert(0, ".")
 import unittest
 from bears.tests.LocalBearTestHelper import LocalBearTestHelper
-from bears.linters.RubyLintBear import RubyLintBear
+from bears.ruby.RubyLintBear import RubyLintBear
 from coalib.settings.Section import Section
