+++ bears/tests/ruby/test_files/rubylint_test2.rb
